remove bespoke hosted embed solution in favour of using unsafe embed block component instead
