+++ dotcom-rendering/src/lib/renderElement.tsx
@@ -19,7 +19,6 @@ import { Figure } from '../components/Figure';
 import { GuideAtomWrapper } from '../components/GuideAtomWrapper.island';
 import { GuVideoBlockComponent } from '../components/GuVideoBlockComponent';
 import { HighlightBlockComponent } from '../components/HighlightBlockComponent';
-import { HostedEmbedBlockComponent } from '../components/HostedEmbedBlockComponent';
 import { ImageBlockComponent } from '../components/ImageBlockComponent';
 import { InstagramBlockComponent } from '../components/InstagramBlockComponent.island';
 import { InteractiveAtom } from '../components/InteractiveAtom';
@@ -320,17 +319,6 @@ export const renderElement = ({
 					);
 				}
 
-				if (format.design === ArticleDesign.HostedArticle) {
-					return (
-						<HostedEmbedBlockComponent
-							html={element.html}
-							alt={element.alt ?? ''}
-							height={element.height}
-							width={element.width}
-						/>
-					);
-				}
-
 				return (
 					<Island priority="feature" defer={{ until: 'visible' }}>
 						<UnsafeEmbedBlockComponent
@@ -344,6 +332,8 @@ export const renderElement = ({
 							source={element.source}
 							sourceDomain={element.sourceDomain}
 							isPinnedPost={isPinnedPost}
+							height={element.height}
+							width={element.width}
 						/>
 					</Island>
 				);
